Move queue package into standalone app
