+++ apps/standalone/package.json
@@ -16,14 +16,17 @@
     "lint": "tsc --noEmit"
   },
   "dependencies": {
+    "@bullstudio/bull-adapter": "workspace:*",
+    "@bullstudio/bullmq-adapter": "workspace:*",
     "@bullstudio/connect-types": "workspace:*",
     "@bullstudio/embedded-core": "workspace:*",
     "@bullstudio/private-router": "workspace:*",
-    "@bullstudio/queue": "workspace:*",
     "@hono/node-server": "^2.0.4",
     "@trpc/server": "^11.4.3",
+    "bull": "^4.16.4",
     "bullmq": "^5.34.8",
-    "hono": "^4.12.23"
+    "hono": "^4.12.23",
+    "ioredis": "^5.4.2"
   },
   "devDependencies": {
     "@bullstudio/typescript-config": "workspace:*",
