+++ apps/standalone/src/queue/detection/provider-detector.ts
