+++ apps/standalone/src/queue/types/connection.types.ts
