+++ apps/standalone/vitest.config.ts
@@ -4,6 +4,14 @@ export default defineConfig({
   test: {
     environment: "node",
     include: ["src/**/*.test.ts"],
+    testTimeout: 15_000,
+    hookTimeout: 15_000,
     globals: false,
+    fileParallelism: false,
+    poolOptions: {
+      threads: {
+        singleThread: true,
+      },
+    },
   },
 });
