--- packages/queue/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@bullstudio/queue",
-  "version": "0.0.1",
-  "private": true,
-  "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "lint": "tsc --noEmit",
-    "typecheck": "tsc --noEmit",
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "@bullstudio/bull-adapter": "workspace:*",
-    "@bullstudio/bullmq-adapter": "workspace:*",
-    "@bullstudio/connect-types": "workspace:*",
-    "bull": "^4.16.4",
-    "bullmq": "^5.34.8",
-    "ioredis": "^5.4.2"
-  },
-  "devDependencies": {
-    "@bullstudio/typescript-config": "workspace:*",
-    "@types/node": "^24.10.1",
-    "typescript": "^5.7.3",
-    "vitest": "^3.2.4"
-  }
-}
